backend_anthesis/requirements.txt
Django>=5.0
djangorestframework
django-filter
psycopg2-binary
gunicorn
django-cors-headers
pytest
pytest-django
pytest-cov
factory-boy
faker
